requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-countries==7.4.2
django-payments==1.0.0
django-phonenumber-field==7.0.0
idna==3.4
instagrapi==1.16.30
phonenumberslite==8.13.1
Pillow==9.3.0
pycryptodomex==3.15.0
pydantic==1.10.2
PySocks==1.7.1
requests==2.28.1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
djangorestframework==3.14.0